+++ views/index.hbs
@@ -1,0 +1,13 @@
+app.use(express.static("public”))
+
+<!DOCTYPE html>
+<html>
+  <head>
+    <title>Do Something</title>
+    <script src="jquery-1.11.3.min.js"></script>
+    <script src="app.js"></script>
+  </head>
+  <body>
+    <h1>Do Something</h1>
+  </body>
+</html>
