+++ views/layout.hbs
@@ -1,0 +1,10 @@
+<!-- <!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8">
+  <title>Do Something</title>
+</head>
+<body>
+  {{{body}}}
+</body>
+</html> -->
